few fixes from begining to understand the structure of the repo.(Errors now cleared)
