+++ packages/core/src/map.ts
@@ -1,46 +1,69 @@
-Map.prototype.set = function <K, V>(
-    this: Map<K, V>,
-    key: K,
-    value: V
-): Map<K, V> {
-    this.set(key, value)
-    return this
+export class Map<K, V> {
+    private map: { key: K; value: V }[] = []
+
+    size(): number {
+        return this.map.length
+    }
+
+    set(key: K, value: V): void {
+        const entry = this.getEntryByKey(key)
+        if (entry) {
+            entry.value = value
+        } else {
+            this.map.push({ key, value })
+        }
+    }
+
+    get(key: K): V | undefined {
+        const entry = this.getEntryByKey(key)
+        return entry ? entry.value : undefined
+    }
+
+    has(key: K): boolean {
+        return this.getEntryByKey(key) !== undefined
+    }
+
+    delete(key: K): boolean {
+        const entryIndex = this.findIndexByKey(key)
+        if (entryIndex !== -1) {
+            const newArray: { key: K; value: V }[] = []
+            for (let i = 0; i < this.map.length; i++) {
+                if (i !== entryIndex) {
+                    newArray.push(this.map[i])
+                }
+            }
+            this.map = newArray
+            return true
+        }
+        return false
+    }
+
+    clear(): void {
+        this.map = []
+    }
+
+    private getEntryByKey(key: K): { key: K; value: V } | undefined {
+        for (let i = 0; i < this.map.length; i++) {
+            const entry = this.map[i]
+            if (this.areKeysEqual(entry.key, key)) {
+                return entry
+            }
+        }
+        return undefined
+    }
+
+    private findIndexByKey(key: K): number {
+        for (let i = 0; i < this.map.length; i++) {
+            const entry = this.map[i]
+            if (this.areKeysEqual(entry.key, key)) {
+                return i
+            }
+        }
+        return -1
+    }
+
+    private areKeysEqual(key1: K, key2: K): boolean {
+        return key1 === key2
+    }
 }
 
-Map.prototype.get = function <K, V>(this: Map<K, V>, key: K): V | undefined {
-    return this.get(key)
-}
-
-Map.prototype.has = function <K, V>(this: Map<K, V>, key: K): boolean {
-    return this.has(key)
-}
-
-Map.prototype.delete = function <K, V>(this: Map<K, V>, key: K): boolean {
-    return this.delete(key)
-}
-
-Map.prototype.clear = function <K, V>(this: Map<K, V>): void {
-    this.clear()
-}
-
-Map.prototype.forEach = function <K, V>(
-    this: Map<K, V>,
-    callbackFn: (value: V, key: K, map: Map<K, V>) => void,
-    thisArg?: any
-): void {
-    this.forEach(callbackFn, thisArg)
-}
-
-Map.prototype.keys = function <K, V>(this: Map<K, V>): IterableIterator<K> {
-    return this.keys()
-}
-
-Map.prototype.values = function <K, V>(this: Map<K, V>): IterableIterator<V> {
-    return this.values()
-}
-
-Map.prototype.entries = function <K, V>(
-    this: Map<K, V>
-): IterableIterator<[K, V]> {
-    return this.entries()
-}
